+++ lib/utils/rk4/models/_models.dart
@@ -1,2 +1,2 @@
-export 'derivative.dart';
-export 'state.dart';
+export 'object_derivative/_object_derivative.dart';
+export 'object_state/object_state.dart';
